Organize #includes, add move function to ship
