+++ Ship.cpp
@@ -4,7 +4,6 @@ Ship::Ship(){
     // We don't know what to set movementSpeed to. 
     // Will wait until we can test on a background of stars to see how it works.
     movementSpeed = 10.0;
-    // Body[0].position = sf::Vector2f(0.f, 50.f);
     Body.setPointCount(36);
 	Body.setPoint(0, sf::Vector2f(0.f, 45.f));
 	Body.setPoint(1, sf::Vector2f(-7.f, 45.f));
@@ -50,14 +49,16 @@ Ship::Ship(){
     setPosition(0, 0);
 }
 
+
+
 void Ship::draw(sf::RenderTarget& target, sf::RenderStates states)const{
     states.transform *= getTransform();
     target.draw(Body, states); 
 }
 
 
+
 void Ship::update(sf::Time dt){
-    sf::Vector2f shipMovement = sf::Vector2f(20.f, 20.f);
 
     if (sf::Keyboard::isKeyPressed(sf::Keyboard::Left)) {
 		rotate(-230.f * dt.asSeconds());
@@ -68,20 +69,25 @@ void Ship::update(sf::Time dt){
 	}
 
 	if (sf::Keyboard::isKeyPressed(sf::Keyboard::R)) {
-		shipMovement = sf::Vector2f(0.f, 0.f);
+		movement = sf::Vector2f(0.f, 0.f);
 		setPosition(0, 0);
-		// coords.setPosition(-(desktop.width / 2.f), -(desktop.height / 2.f) + 50);
 	}
 
 	if (sf::Keyboard::isKeyPressed(sf::Keyboard::Up)) {
-		shipMovement += movementSpeed * dt.asSeconds() * sf::Vector2f(-sin(getRotation() * (3.1415 / 180)), 
+		movement += movementSpeed * dt.asSeconds() * sf::Vector2f(-sin(getRotation() * (3.1415 / 180)), 
 			cos(getRotation() * (3.1415 / 180))); 
 	}
 
 	if (sf::Keyboard::isKeyPressed(sf::Keyboard::Down)) {
-		shipMovement -= movementSpeed * dt.asSeconds() * sf::Vector2f(-sin(getRotation() * (3.1415 / 180)), 
+		movement -= movementSpeed * dt.asSeconds() * sf::Vector2f(-sin(getRotation() * (3.1415 / 180)), 
 			cos(getRotation() * (3.1415 / 180)));
 	}
 
 
+}
+
+
+
+void Ship::move(sf::Time dt){
+    sf::Transformable::move(movement * dt.asSeconds());
 }
